refactor(mobile): lift the terminal attachment lifecycle into a hook
